refactor(tests): move demo-related tests to test_demo.py
