+++ everyvoice/tests/test_cli.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 
-import enum
 import json
 import os
 import subprocess
@@ -12,7 +11,6 @@ from typing import Any
 from unittest import TestCase, mock
 
 import jsonschema
-import typer
 import yaml
 from packaging.version import Version
 from pydantic import ValidationError
@@ -20,8 +18,6 @@ from pytest import main
 from pytorch_lightning import Trainer
 from typer.testing import CliRunner
 from yaml import CLoader as Loader
-
-import everyvoice.tests.model_stubs
 
 # required for `./run_tests.py cli` to work, otherwise test_inspect_checkpoint
 # fails with an Intel MKL FATAL ERROR saying it cannot load libtorch_cpu.so
@@ -31,7 +27,6 @@ from everyvoice._version import VERSION
 from everyvoice.base_cli.helpers import save_configuration_to_log_dir
 from everyvoice.cli import SCHEMAS_TO_OUTPUT, app
 from everyvoice.config.shared_types import ContactInformation
-from everyvoice.demo.app import create_demo_app
 from everyvoice.model.feature_prediction.FastSpeech2_lightning.fs2.config import (
     FastSpeech2Config,
 )
@@ -45,9 +40,11 @@ from everyvoice.model.feature_prediction.FastSpeech2_lightning.fs2.type_definiti
 from everyvoice.model.vocoder.HiFiGAN_iSTFT_lightning.hfgl.config import HiFiGANConfig
 from everyvoice.model.vocoder.HiFiGAN_iSTFT_lightning.hfgl.model import HiFiGAN
 from everyvoice.tests.stubs import (
+    TEST_DATA_DIR,
     capture_logs,
     capture_stdout,
     flatten_log,
+    mock_function_placeholder,
     temp_chdir,
 )
 from everyvoice.wizard import (
@@ -63,23 +60,10 @@ def major_minor(version):
     return f"{v.major}.{v.minor}"
 
 
-def mock_function_placeholder(*args, **kwargs):
-    """
-    Mock function to replace any function that we are not testing.
-    """
-    print("mock_fuction_placeholder called with args:", args, "and kwargs:", kwargs)
-
-
-def mock_function_placeholder2(**kwargs):
-    """
-    Mock function to replace any function that we are not testing.
-    """
-    print("mock_fuction_placeholder2 called with kwargs:", kwargs)
+CONFIG_DIR = TEST_DATA_DIR / "relative" / "config"
 
 
 class CLITest(TestCase):
-    data_dir = Path(__file__).parent / "data"
-    config_dir = Path(__file__).parent / "data" / "relative" / "config"
     class_tmp_dir_obj: Any
     class_tmp_dir: str
     dummy_fp_path: Path
@@ -133,12 +117,12 @@ class CLITest(TestCase):
             random.seed(10719787423044995460)
             vocoder = HiFiGAN(
                 HiFiGANConfig.load_config_from_path(
-                    cls.config_dir / f"{SPEC_TO_WAV_CONFIG_FILENAME_PREFIX}.yaml"
+                    CONFIG_DIR / f"{SPEC_TO_WAV_CONFIG_FILENAME_PREFIX}.yaml"
                 )
             )
             spec_model = FastSpeech2(
                 FastSpeech2Config.load_config_from_path(
-                    cls.config_dir / f"{TEXT_TO_SPEC_CONFIG_FILENAME_PREFIX}.yaml"
+                    CONFIG_DIR / f"{TEXT_TO_SPEC_CONFIG_FILENAME_PREFIX}.yaml"
                 ),
                 lang2id={"default": 0},
                 speaker2id={"default": 0},
@@ -349,9 +333,9 @@ class CLITest(TestCase):
             [
                 "evaluate",
                 "-f",
-                self.data_dir / "LJ010-0008.wav",
+                TEST_DATA_DIR / "LJ010-0008.wav",
                 "-r",
-                self.data_dir / "lj" / "wavs" / "LJ050-0269.wav",
+                TEST_DATA_DIR / "lj" / "wavs" / "LJ050-0269.wav",
             ],
         )
         assert result.exit_code == 0
@@ -365,9 +349,9 @@ class CLITest(TestCase):
             [
                 "evaluate",
                 "-d",
-                self.data_dir / "lj" / "wavs",
+                TEST_DATA_DIR / "lj" / "wavs",
                 "-r",
-                self.data_dir / "LJ010-0008.wav",
+                TEST_DATA_DIR / "LJ010-0008.wav",
             ],
         )
         assert dir_result.exit_code == 0
@@ -377,13 +361,13 @@ class CLITest(TestCase):
             dir_result.stdout,
             "should report metrics in terms of averages",
         )
-        evaluation_output = self.data_dir / "lj" / "wavs" / "evaluation.json"
+        evaluation_output = TEST_DATA_DIR / "lj" / "wavs" / "evaluation.json"
         assert evaluation_output.exists(), "should print results to a file"
         evaluation_output.unlink()
 
     def test_old_inspect_checkpoint(self):
         result = self.runner.invoke(
-            app, ["inspect-checkpoint", str(self.data_dir / "test.ckpt")]
+            app, ["inspect-checkpoint", str(TEST_DATA_DIR / "test.ckpt")]
         )
         assert result.exit_code == 0
         self.assertIn(
@@ -397,7 +381,7 @@ class CLITest(TestCase):
 
     def test_inspect_checkpoint(self):
         result = self.runner.invoke(
-            app, ["checkpoint", "inspect", str(self.data_dir / "test.ckpt")]
+            app, ["checkpoint", "inspect", str(TEST_DATA_DIR / "test.ckpt")]
         )
         assert 'global_step": 52256' in result.stdout
         self.assertIn(
@@ -458,7 +442,7 @@ class CLITest(TestCase):
                 [
                     "preprocess",
                     "text-to-spec",
-                    str(self.config_dir / "everyvoice-spec-to-wav.yaml"),
+                    str(CONFIG_DIR / "everyvoice-spec-to-wav.yaml"),
                 ],
             )
             assert result.exit_code == 1
@@ -499,101 +483,15 @@ class CLITest(TestCase):
         self.assertNotIn(b"shared_types", result.stderr, msg.format("shared_types.py"))
         self.assertNotIn(b"pydantic", result.stderr, msg.format("pydantic"))
 
-    def test_demo_with_bad_args(self):
-        # No checkpoint → help message
-        result = self.runner.invoke(app, ["demo"])
-        # Exit code for no-arg-is-help is 0 with click<=8.1.8 and typer<=0.23.2,
-        # 2 if either is more recent
-        assert result.exit_code in (0, 2)
-        # the runner calls "root" instead of "everyvoice"
-        assert "Usage: root demo [OPTIONS] CHECKPOINT" in flatten_log(result.output)
-
-        # Invalid --output-format value
-        result = self.runner.invoke(
-            app,
-            [
-                "demo",
-                os.devnull,
-                "--vocoder",
-                os.devnull,
-                "--output-format",
-                "not-a-format",
-            ],
-        )
-        assert result.exit_code != 0
-        assert "Invalid value" in flatten_log(result.output)
-
-    EMPTY_DEMO_ARGS: dict[str, Any] = {
-        "languages": [],
-        "speakers": [],
-        "output_dir": None,
-        "accelerator": None,
-    }
-
-    def test_create_demo_app_with_errors(self):
-        # outputs is the first thing to get checked, because it can be done as
-        # a quick check before loading any models.
-        with self.assertRaises(ValueError) as cm:
-            create_demo_app(
-                text_to_spec_model_path=None,
-                spec_to_wav_model_path=None,
-                **self.EMPTY_DEMO_ARGS,  # type: ignore[arg-type]
-                outputs=[],
-            )
-        assert "Empty outputs list" in str(cm.exception)
-
-        class WrongEnum(str, enum.Enum):
-            foo = "foo"
-
-        for outputs in (["wav", WrongEnum.foo], ["textgrid", "foo"]):
-            with self.assertRaises(ValueError) as cm:
-                create_demo_app(
-                    text_to_spec_model_path=None,
-                    spec_to_wav_model_path=None,
-                    **self.EMPTY_DEMO_ARGS,  # type: ignore[arg-type]
-                    outputs=outputs,
-                )
-            assert "Unknown output format 'foo'" in str(cm.exception)
-
-    def test_demo_with_bad_models(self) -> None:
-        devnull = Path(os.devnull)
-        with self.assertRaises(ValueError) as cm:
-            create_demo_app(devnull, devnull, **self.EMPTY_DEMO_ARGS, outputs=["wav"])  # type: ignore[arg-type]
-        assert "It does not appear to be a valid checkpoint" in str(cm.exception)
-
-        with self.assertRaises(ValueError) as cm:
-            create_demo_app(
-                devnull,
-                self.data_dir / "test.ckpt",
-                **self.EMPTY_DEMO_ARGS,  # type: ignore[arg-type]
-                outputs=["wav"],
-            )
-        assert "maybe it's not actually a HiFiGAN model" in str(cm.exception)
-
-    def test_demo_with_wrong_models(self) -> None:
-        fp_path, vocoder_path = self.get_dummy_models()
-        with self.assertRaises(ValueError) as cm:
-            create_demo_app(fp_path, fp_path, **self.EMPTY_DEMO_ARGS, outputs=["wav"])  # type: ignore[arg-type]
-        assert "maybe it's not actually a HiFiGAN model" in str(cm.exception)
-
-        with self.assertRaises(ValueError) as cm:
-            create_demo_app(
-                vocoder_path,
-                vocoder_path,
-                **self.EMPTY_DEMO_ARGS,  # type: ignore[arg-type]
-                outputs=["wav"],
-            )
-        assert "maybe it's not actually an fs2 model" in str(cm.exception)
-
     def test_g2p(self):
         result = self.runner.invoke(
             app,
             [
                 "g2p",
                 "abc",
-                str(self.data_dir / Path("text.txt")),
+                str(TEST_DATA_DIR / "text.txt"),
                 "--config",
-                str(self.config_dir / Path("everyvoice-shared-text.yaml")),
+                str(CONFIG_DIR / "everyvoice-shared-text.yaml"),
             ],
         )
 
@@ -692,466 +590,6 @@ class CLITest(TestCase):
                 # print(result.output)
                 assert result.exit_code != 0
                 assert "No speakers found" in flatten_log(result.output)
-
-
-class TestDemo(TestCase):
-    def mock_create_demo_app(self, *_args, **_kwargs):
-        class MockCreateDemoApp:
-            def launch(self, *_args, **_kwargs):
-                print(f"  - Launch Port: {_kwargs['server_port']}")
-                print(f"  - Launch Share: {_kwargs['share']}")
-                print(f"  - Launch Server Name: {_kwargs['server_name']}")
-                if "config_file" in _kwargs:
-                    print(f"  - Config File: {_kwargs['config_file']}")
-                else:
-                    print("  - Config File: None")
-
-        return MockCreateDemoApp()
-
-    def test_create_demo_app(self):
-        with tempfile.TemporaryDirectory() as tmpdir_str:
-            tmpdir = Path(tmpdir_str)
-            # This test is just to make sure that the demo app can be created with parameters for gradio
-            # and that it doesn't crash.
-            _, vocoder_path = everyvoice.tests.model_stubs.get_stubbed_vocoder(
-                tmpdir / "vocoder"
-            )
-            _, spec_model_path = everyvoice.tests.model_stubs.get_stubbed_model(
-                tmpdir / "spec_model"
-            )
-            # This test is just to make sure that the demo app params are passed correctly
-            port = 7000
-            ip = "123.456.78.90"
-            # with mock.patch(
-            #    "everyvoice.demo.app.create_demo_app",
-            #    side_effect=self.mock_create_demo_app,
-            # ):
-            with (
-                mock.patch(
-                    "everyvoice.cli._peek_model_class",
-                    return_value="FastSpeech2",
-                ),
-                mock.patch(
-                    "everyvoice.demo.app.load_model_from_checkpoint",
-                    side_effect=self.mock_demo_load_model_from_checkpoint,
-                ),
-                mock.patch(
-                    "everyvoice.base_cli.helpers.inference_base_command",
-                    side_effect=mock_function_placeholder,
-                ),
-                mock.patch(
-                    "everyvoice.demo.app.synthesize_audio",
-                    side_effect=mock_function_placeholder2,
-                ),
-                mock.patch(
-                    "gradio.Blocks.launch",
-                    return_value="Launching gradio app blocks",
-                    side_effect=mock_function_placeholder2,
-                ),
-            ):
-                result = CliRunner().invoke(
-                    app,
-                    [
-                        "demo",
-                        str(spec_model_path),
-                        "--vocoder",
-                        str(vocoder_path),
-                        "--port",
-                        port,
-                        "--share",
-                        "--server-name",
-                        ip,  # Mock IP address
-                    ],
-                )
-            assert result.exit_code == 0
-            assert f"Port: {port}" in flatten_log(result.output)
-            assert "Share: True" in flatten_log(result.output)
-            assert f"Server Name: {ip}" in flatten_log(result.output)
-
-    def mock_demo_load_model_from_checkpoint(
-        *_arg, **kwargs
-    ) -> tuple:  # [FastSpeech2, torch.nn.Module, dict, torch.device]:
-        print(
-            "mock_demo_load_model_from_checkpoint called with args:",
-            _arg,
-            "and kwargs:",
-            kwargs,
-        )
-
-        class model:
-            from types import SimpleNamespace
-
-            lang2id = {"default": 0}
-            speaker2id = {"default": 0}
-            model_data = {"use_global_style_token_module": False}
-            config_data = {"model": SimpleNamespace(**model_data)}
-            config = SimpleNamespace(
-                **config_data,
-            )
-
-        return model, {}, {}, "cpu"  # Mock return values for the test
-
-    def test_create_demo_app_with_ui_config_file(self) -> None:
-        with tempfile.TemporaryDirectory() as tmpdir_str:
-            tmpdir = Path(tmpdir_str)
-            # This test is just to make sure that the demo app can be created with parameters for gradio
-            # and that it doesn't crash.
-            _, vocoder_path = everyvoice.tests.model_stubs.get_stubbed_vocoder(
-                tmpdir / "vocoder"
-            )
-            _, spec_model_path = everyvoice.tests.model_stubs.get_stubbed_model(
-                tmpdir / "spec_model"
-            )
-
-            # Create a dummy app config file
-            config: dict = {
-                "app_title": "Test App",
-                "app_description": "This is a test app description.",
-                "app_instructions": "These are test app instructions.",
-                "speakers": {
-                    "default": "Person A",
-                },
-                "languages": {
-                    "default": "English",
-                },
-                "input_text_label": "Input Text",
-                "duration_multiplier_label": "Duration Multiplier",
-                "language_label": "Language",
-                "speaker_label": "Speaker",
-                "output_format_label": "Output Format",
-                "synthesize_label": "Synthesize",
-                "file_output_label": "File Output",
-            }
-            config_file = tmpdir / "demo_config.json"
-            with config_file.open("w", encoding="utf8") as f:
-                json.dump(config, f)
-            allowlist_file = tmpdir / "allowlist.txt"
-            with allowlist_file.open("w", encoding="utf8") as f:
-                f.write("hey\nyes\nword")
-            # This test is just to make sure that the demo app params are passed correctly
-            port = "7000"
-            ip = "123.456.78.90"
-
-            with (
-                mock.patch(
-                    "everyvoice.cli._peek_model_class",
-                    return_value="FastSpeech2",
-                ),
-                mock.patch(
-                    "everyvoice.demo.app.load_model_from_checkpoint",
-                    side_effect=self.mock_demo_load_model_from_checkpoint,
-                ),
-                mock.patch(
-                    "everyvoice.base_cli.helpers.inference_base_command",
-                    side_effect=mock_function_placeholder,
-                ),
-                mock.patch(
-                    "everyvoice.demo.app.synthesize_audio",
-                    side_effect=mock_function_placeholder2,
-                ),
-                mock.patch(
-                    "gradio.Blocks.launch",
-                    return_value="Launching gradio app blocks",
-                    side_effect=mock_function_placeholder2,
-                ),
-            ):
-                result = CliRunner().invoke(
-                    app,
-                    [
-                        "demo",
-                        str(spec_model_path),
-                        "--vocoder",
-                        str(vocoder_path),
-                        "--port",
-                        port,
-                        "--server-name",
-                        ip,  # Mock IP address
-                        "--ui-config-file",
-                        str(config_file),
-                        "--speaker",
-                        "default",
-                        "--language",
-                        "default",
-                        "--allowlist",
-                        allowlist_file,
-                    ],
-                )
-                # print(result.output, result.exit_code)  # Debug output
-
-            assert result.exit_code == 0
-            self.assertIn(
-                f"Using speakers from app config JSON: [('{config['speakers']['default']}', 'default')]",
-                result.output,
-            )
-            self.assertIn(
-                f"Using languages from app config JSON: [('{config['languages']['default']}', 'default')]",
-                result.output,
-            )
-
-            self.assertIn(
-                f"Using app title from app config JSON: {config['app_title']}",
-                result.output,
-            )
-
-    def test_create_demo_app_with_malformed_ui_config_file(self):
-        with tempfile.TemporaryDirectory() as tmpdir_str:
-            tmpdir = Path(tmpdir_str)
-            # This test is just to make sure that the demo app can be created with parameters for gradio
-            # and that it doesn't crash.
-            _, vocoder_path = everyvoice.tests.model_stubs.get_stubbed_vocoder(
-                tmpdir / "vocoder"
-            )
-            _, spec_model_path = everyvoice.tests.model_stubs.get_stubbed_model(
-                tmpdir / "spec_model"
-            )
-            # Create a malformed app config file (missing closing brace)
-            config = """{
-                "app_title": "Test App",
-                "app_description": "This is a test app description.",
-                "app_instructions": "These are test app instructions.",
-                "input_text_label": "Input Text",
-                "duration_multiplier_label": "Duration Multiplier",
-                "language_label": "Language",
-                "speaker_label": "Speaker",
-                "output_format_label": "Output Format",
-                "synthesize_label": "Synthesize",
-                "file_output_label": "File Output",
-            """
-            config_file = tmpdir / "malformed_demo_config.json"
-            with config_file.open("w", encoding="utf8") as f:
-                f.write(config)
-            # This test is just to make sure that the demo app params are passed correctly
-            port = "7000"
-            ip = "123.456.78.90"
-
-            with (
-                mock.patch(
-                    "everyvoice.cli._peek_model_class",
-                    return_value="FastSpeech2",
-                ),
-                mock.patch(
-                    "everyvoice.demo.app.load_model_from_checkpoint",
-                    side_effect=self.mock_demo_load_model_from_checkpoint,
-                ),
-                mock.patch(
-                    "everyvoice.base_cli.helpers.inference_base_command",
-                    side_effect=mock_function_placeholder,
-                ),
-                mock.patch(
-                    "everyvoice.demo.app.synthesize_audio",
-                    side_effect=mock_function_placeholder2,
-                ),
-                mock.patch(
-                    "gradio.Blocks.launch",
-                    return_value="Launching gradio app blocks",
-                    side_effect=mock_function_placeholder2,
-                ),
-            ):
-                result = CliRunner().invoke(
-                    app,
-                    [
-                        "demo",
-                        str(spec_model_path),
-                        "--vocoder",
-                        str(vocoder_path),
-                        "--port",
-                        port,
-                        "--server-name",
-                        ip,  # Mock IP address
-                        "--ui-config-file",
-                        str(config_file),
-                        "--speaker",
-                        "default",
-                        "--language",
-                        "default",
-                    ],
-                )
-                # print(result.output, result.exit_code)  # Debug output
-            assert result.exit_code != 0
-            self.assertRegex(
-                result.output, r"(?s)Your config file.*malformed.*has.*errors"
-            )
-
-    # unit test for error handling in load_app_ui_labels
-    def test_create_demo_load_app_ui_labels_errors(self):
-        from everyvoice.demo.app import load_app_ui_labels
-
-        # Create a dummy app config file
-        config_bad_speaker = {
-            "app_title": "Test App",
-            "speakers": {
-                "unknown": "Person A",
-            },
-            "languages": {
-                "default": "English",
-            },
-        }
-        config_bad_language = {
-            "app_title": "Test App",
-            "speakers": {
-                "default": "Person A",
-            },
-            "languages": {
-                "unknown": "English",
-            },
-        }
-
-        with self.assertRaises(typer.BadParameter) as cm:
-            load_app_ui_labels(
-                config_bad_language,
-                ["all"],
-                ["all"],
-                ["default"],
-                ["default"],
-            )
-        self.assertIn(
-            "The 'languages' key in the app config JSON does not match the languages provided.",
-            str(cm.exception),
-        )
-        with self.assertRaises(typer.BadParameter) as cm:
-            load_app_ui_labels(
-                config_bad_speaker,
-                ["all"],
-                ["all"],
-                ["default"],
-                ["default"],
-            )
-        self.assertIn(
-            "The 'speakers' key in the app config JSON does not match the speakers provided.",
-            str(cm.exception),
-        )
-        with self.assertRaises(typer.BadParameter) as cm:
-            load_app_ui_labels(
-                config_bad_speaker,
-                ["default"],
-                ["unknown"],
-                ["default"],
-                ["default"],
-            )
-
-        self.assertIn(
-            "Language option has been activated, but valid languages have not been provided. The model has been trained in ['default'] languages. Please select either 'all' or at least some of them.",
-            str(cm.exception),
-        )
-        with self.assertRaises(typer.BadParameter) as cm:
-            load_app_ui_labels(
-                config_bad_speaker,
-                ["unknown"],
-                ["default"],
-                ["default"],
-                ["default"],
-            )
-        self.assertIn(
-            "Speaker option has been activated, but valid speakers have not been provided. The model has been trained with ['default'] speakers. Please select either 'all' or at least some of them.",
-            str(cm.exception),
-        )
-
-    def test_demo_dispatch_styletts2_rejects_vocoder_flag(self):
-        """Passing --vocoder with a StyleTTS2 checkpoint should produce a clear error."""
-        with tempfile.TemporaryDirectory() as tmpdir_str:
-            import torch
-
-            tmpdir = Path(tmpdir_str)
-            fake_ckpt = tmpdir / "styletts2.ckpt"
-            torch.save(
-                {
-                    "model_info": {"name": "StyleTTS2Module"},
-                    "hyper_parameters": {"mode": "second", "config": {}},
-                    "state_dict": {},
-                },
-                fake_ckpt,
-            )
-            fake_vocoder = tmpdir / "hifigan.ckpt"
-            fake_vocoder.touch()
-
-            result = CliRunner().invoke(
-                app,
-                [
-                    "demo",
-                    str(fake_ckpt),
-                    "--vocoder",
-                    str(fake_vocoder),
-                    "--ref-speaker",
-                    f"Eric={fake_ckpt}",  # reuse fake_ckpt as a dummy audio file
-                ],
-            )
-            assert result.exit_code != 0
-            assert "StyleTTS2 does not use a separate vocoder" in flatten_log(
-                result.output
-            )
-
-    def test_demo_dispatch_fs2_requires_vocoder(self):
-        """Invoking demo with a FastSpeech2 checkpoint but no --vocoder should error."""
-        with tempfile.TemporaryDirectory() as tmpdir_str:
-            tmpdir = Path(tmpdir_str)
-            _, spec_model_path = everyvoice.tests.model_stubs.get_stubbed_model(
-                tmpdir / "spec_model"
-            )
-
-            with mock.patch(
-                "everyvoice.cli._peek_model_class",
-                return_value="FastSpeech2",
-            ):
-                result = CliRunner().invoke(
-                    app,
-                    ["demo", str(spec_model_path)],
-                )
-            assert result.exit_code != 0
-            assert "FastSpeech2 requires a vocoder checkpoint" in flatten_log(
-                result.output
-            )
-
-    def test_demo_dispatch_fs2_rejects_ref_speaker(self):
-        """Passing --ref-speaker with a FastSpeech2 checkpoint should produce a clear error."""
-        with tempfile.TemporaryDirectory() as tmpdir_str:
-            tmpdir = Path(tmpdir_str)
-            _, vocoder_path = everyvoice.tests.model_stubs.get_stubbed_vocoder(
-                tmpdir / "vocoder"
-            )
-            _, spec_model_path = everyvoice.tests.model_stubs.get_stubbed_model(
-                tmpdir / "spec_model"
-            )
-
-            with mock.patch(
-                "everyvoice.cli._peek_model_class",
-                return_value="FastSpeech2",
-            ):
-                result = CliRunner().invoke(
-                    app,
-                    [
-                        "demo",
-                        str(spec_model_path),
-                        "--vocoder",
-                        str(vocoder_path),
-                        "--ref-speaker",
-                        f"Eric={spec_model_path}",
-                    ],
-                )
-            assert result.exit_code != 0
-            assert "--ref-speaker is only used with StyleTTS2" in flatten_log(
-                result.output
-            )
-
-    def test_demo_dispatch_vocoder_checkpoint_as_primary(self):
-        """Passing a HiFiGAN checkpoint as the primary CHECKPOINT should give a helpful error."""
-        with tempfile.TemporaryDirectory() as tmpdir_str:
-            import torch
-
-            tmpdir = Path(tmpdir_str)
-            fake_vocoder_ckpt = tmpdir / "hifigan.ckpt"
-            torch.save(
-                {"model_info": {"name": "HiFiGAN"}, "state_dict": {}},
-                fake_vocoder_ckpt,
-            )
-
-            result = CliRunner().invoke(
-                app,
-                ["demo", str(fake_vocoder_ckpt)],
-            )
-            assert result.exit_code != 0
-            assert "appears to be a standalone vocoder checkpoint" in flatten_log(
-                result.output
-            )
 
 
 class TestBaseCLIHelper(TestCase):
